mapariesgos/requirements.txt
django==3.2.15
djangorestframework==3.13.1
pillow
mysqlclient